+++ threedi_modelchecker/config.py
@@ -29,6 +29,7 @@ from .checks.other import ConnectionNodesLength
 from .checks.other import CrossSectionLocationCheck
 from .checks.other import LinestringLocationCheck
 from .checks.other import OpenChannelsWithNestedNewton
+from .checks.other import SpatialIndexCheck
 from .checks.other import Use0DFlowCheck
 from .checks.timeseries import TimeseriesIncreasingCheck
 from .checks.timeseries import TimeseriesRowCheck
@@ -533,6 +534,11 @@ CHECKS += [
             models.ConnectionNode.the_geom_linestring != None
         ),
         message="The 'the_geom_linestring' column of v2_connection_nodes must be NULL",
+    )
+]
+CHECKS += [
+    SpatialIndexCheck(
+        error_code=207, column=models.ConnectionNode.the_geom, level=CheckLevel.WARNING
     )
 ]